requirements.txt
tensorflow
keras
jupyterlab
sklearn
pandas
Werkzeug
Flask
numpy
Keras
gevent
pillow
h5py